https://github.com/GreenInfo-Network/east-coast-greenway/issues/29 rename poiInfoActions.js
